1-20 для встречи со всеми! людьми.

+++ Armlet_fw/SharedAtlLogic/emotions.h
@@ -152,7 +152,7 @@ extern Emotion_t emotions[];
 typedef struct Reason {
     const char* name;
     int weight;
-    int age;
+    signed long age;
     int eID;
 } Reason_t;
